+++ pkg/sass_api/CHANGELOG.md
@@ -4,6 +4,8 @@
   it now has a more generic `plainCss` argument which tells it to parse the
   selector in plain CSS mode.
 
+* Rename `SelectorList.resolveParentSelectors` to `SelectorList.nestWithin`.
+
 ## 9.5.0
 
 * No user-visible changes.